backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.4
weasyprint==61.2
pyyaml==6.0.1
jinja2==3.1.3
python-dotenv==1.0.0
beautifulsoup4==4.12.3
